requirements.txt
asgiref==3.8.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
channels==4.2.2
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
crispy-bootstrap5==2024.10
dash==2.18.2
dash-bootstrap-components==1.7.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
defusedxml==0.7.1
diff-match-patch==20241021
dj-database-url==2.3.0
Django==5.1.7
django-appconf==1.1.0
django-crispy-forms==2.3
django-environ==0.12.0
django-filter==25.1
django-htmx==1.23.0
django-import-export==4.3.7
django-plotly-dash==2.4.5
django-select2==8.4.0
django-tables2==2.7.5
djangorestframework==3.15.2
dpd-static-support==0.0.5
dpd_components==0.1.0
et_xmlfile==2.0.0
Faker==37.1.0
Flask==3.0.3
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.69.2
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
narwhals==1.33.0
nest-asyncio==1.6.0
numpy==2.2.4
odfpy==1.4.1
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.2.1
plotly==6.0.1
proto-plus==1.26.1
protobuf==6.30.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
reportlab==4.3.1
requests==2.32.3
retrying==1.3.4
rsa==4.9
setuptools==78.1.0
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
tabulate==0.9.0
typing_extensions==4.13.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.0.6
whitenoise==6.9.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.21.0
